requirements.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
redis==7.0.1
python-dotenv==1.2.1
rich==13.9.4
pytest==8.4.2
